Modularize package tooling and bandwidth limiter
