+++ crates/bandwidth/src/limiter/core.rs
@@ -1,0 +1,187 @@
+use super::{
+    MICROS_PER_SECOND, MIN_WRITE_MAX, MINIMUM_SLEEP_MICROS, duration_from_microseconds, sleep_for,
+};
+use std::num::NonZeroU64;
+use std::time::{Duration, Instant};
+
+fn calculate_write_max(limit: NonZeroU64, burst: Option<NonZeroU64>) -> usize {
+    let kib = if limit.get() < 1024 {
+        1
+    } else {
+        limit.get() / 1024
+    };
+
+    let base_write_max = u128::from(kib)
+        .saturating_mul(128)
+        .max(MIN_WRITE_MAX as u128);
+    let mut write_max = base_write_max.min(usize::MAX as u128) as usize;
+
+    if let Some(burst) = burst {
+        let burst = burst.get().min(usize::MAX as u64);
+        write_max = usize::try_from(burst)
+            .unwrap_or(usize::MAX)
+            .max(MIN_WRITE_MAX)
+            .max(1);
+    }
+
+    write_max.max(MIN_WRITE_MAX)
+}
+
+/// Token-bucket style limiter that mirrors upstream rsync's pacing rules.
+#[doc(alias = "--bwlimit")]
+#[derive(Clone, Debug)]
+pub struct BandwidthLimiter {
+    limit_bytes: NonZeroU64,
+    write_max: usize,
+    burst_bytes: Option<NonZeroU64>,
+    total_written: u128,
+    last_instant: Option<Instant>,
+    simulated_elapsed_us: u128,
+}
+
+impl BandwidthLimiter {
+    /// Constructs a new limiter from the supplied byte-per-second rate.
+    #[must_use]
+    pub fn new(limit: NonZeroU64) -> Self {
+        Self::with_burst(limit, None)
+    }
+
+    /// Constructs a new limiter from the supplied rate and optional burst size.
+    #[must_use]
+    pub fn with_burst(limit: NonZeroU64, burst: Option<NonZeroU64>) -> Self {
+        let write_max = calculate_write_max(limit, burst);
+
+        Self {
+            limit_bytes: limit,
+            write_max,
+            burst_bytes: burst,
+            total_written: 0,
+            last_instant: None,
+            simulated_elapsed_us: 0,
+        }
+    }
+
+    /// Updates the limiter so a new byte-per-second limit takes effect.
+    pub fn update_limit(&mut self, limit: NonZeroU64) {
+        self.update_configuration(limit, self.burst_bytes);
+    }
+
+    /// Updates the limiter so both the rate and burst configuration take effect.
+    #[doc(alias = "--bwlimit")]
+    pub fn update_configuration(&mut self, limit: NonZeroU64, burst: Option<NonZeroU64>) {
+        let write_max = calculate_write_max(limit, burst);
+
+        self.limit_bytes = limit;
+        self.write_max = write_max;
+        self.burst_bytes = burst;
+        self.total_written = 0;
+        self.last_instant = None;
+        self.simulated_elapsed_us = 0;
+    }
+
+    /// Resets the limiter while keeping the current configuration.
+    pub fn reset(&mut self) {
+        self.total_written = 0;
+        self.last_instant = None;
+        self.simulated_elapsed_us = 0;
+    }
+
+    #[inline]
+    fn clamp_debt_to_burst(&mut self) {
+        if let Some(burst) = self.burst_bytes {
+            let limit = u128::from(burst.get());
+            self.total_written = self.total_written.min(limit);
+        }
+    }
+
+    /// Returns the configured limit in bytes per second.
+    #[must_use]
+    pub const fn limit_bytes(&self) -> NonZeroU64 {
+        self.limit_bytes
+    }
+
+    /// Returns the configured burst size in bytes, if any.
+    #[must_use]
+    pub const fn burst_bytes(&self) -> Option<NonZeroU64> {
+        self.burst_bytes
+    }
+
+    /// Returns the maximum chunk size the limiter schedules before sleeping.
+    #[must_use]
+    pub const fn write_max_bytes(&self) -> usize {
+        self.write_max
+    }
+
+    /// Returns the maximum chunk size that should be written before sleeping.
+    #[must_use]
+    pub fn recommended_read_size(&self, buffer_len: usize) -> usize {
+        let limit = self.write_max.max(1);
+        buffer_len.min(limit)
+    }
+
+    /// Records a completed write and sleeps if the limiter accumulated debt.
+    pub fn register(&mut self, bytes: usize) -> super::LimiterSleep {
+        if bytes == 0 {
+            return super::LimiterSleep::default();
+        }
+
+        self.total_written = self.total_written.saturating_add(bytes as u128);
+        self.clamp_debt_to_burst();
+
+        let start = Instant::now();
+        let bytes_per_second = u128::from(self.limit_bytes.get());
+
+        let mut elapsed_us = self.simulated_elapsed_us;
+        if let Some(previous) = self.last_instant {
+            let elapsed = start.duration_since(previous);
+            let measured = elapsed.as_micros().min(u128::from(u64::MAX));
+            elapsed_us = elapsed_us.saturating_add(measured);
+        }
+        self.simulated_elapsed_us = 0;
+        if elapsed_us > 0 {
+            let allowed = elapsed_us.saturating_mul(bytes_per_second) / MICROS_PER_SECOND;
+            if allowed >= self.total_written {
+                self.total_written = 0;
+            } else {
+                self.total_written -= allowed;
+            }
+        }
+
+        self.clamp_debt_to_burst();
+
+        let sleep_us = self.total_written.saturating_mul(MICROS_PER_SECOND) / bytes_per_second;
+
+        if sleep_us < MINIMUM_SLEEP_MICROS {
+            self.last_instant = Some(start);
+            return super::LimiterSleep::default();
+        }
+
+        let requested = duration_from_microseconds(sleep_us);
+        if !requested.is_zero() {
+            sleep_for(requested);
+        }
+
+        let end = Instant::now();
+        let elapsed_us = end
+            .checked_duration_since(start)
+            .map(|duration| duration.as_micros().min(u128::from(u64::MAX)))
+            .unwrap_or(0);
+        if sleep_us > elapsed_us {
+            self.simulated_elapsed_us = sleep_us - elapsed_us;
+        }
+        let remaining_us = sleep_us.saturating_sub(elapsed_us);
+        let leftover = remaining_us.saturating_mul(bytes_per_second) / MICROS_PER_SECOND;
+
+        self.total_written = leftover;
+        self.clamp_debt_to_burst();
+        self.last_instant = Some(end);
+        let actual = Duration::from_micros(elapsed_us as u64);
+        super::LimiterSleep::new(requested, actual)
+    }
+
+    #[cfg(any(test, feature = "test-support"))]
+    #[allow(dead_code)]
+    pub(crate) fn accumulated_debt_for_testing(&self) -> u128 {
+        self.total_written
+    }
+}
